Fixed build on 64bit, FIX #4

+++ source/ColumnListView/CLVEasyItem.cpp
@@ -63,7 +63,7 @@ CLVEasyItem::~CLVEasyItem()
 	int num_columns = m_column_types.CountItems();
 	for(int column = 0; column < num_columns; column++)
 	{
-		int32 type = (int32)m_column_types.ItemAt(column);
+		intptr_t type = (intptr_t)m_column_types.ItemAt(column);
 		bool bitmap_is_copy = false;
 		if(type & CLVColFlagBitmapIsCopy)
 			bitmap_is_copy = true;
@@ -94,7 +94,7 @@ void CLVEasyItem::PrepListsForSet(int column_index)
 	if(delete_old)
 	{
 		//Column content exists already so delete the old entries
-		int32 old_type = (int32)m_column_types.ItemAt(column_index);
+		intptr_t old_type = (intptr_t)m_column_types.ItemAt(column_index);
 		bool bitmap_is_copy = false;
 		if(old_type & CLVColFlagBitmapIsCopy)
 			bitmap_is_copy = true;
@@ -121,7 +121,7 @@ void CLVEasyItem::SetColumnContent(int column_index, const char *text, bool trun
 	((BRect**)m_cached_rects.Items())[column_index]->Set(-1,-1,-1,-1);
 	if(text == NULL || text[0] == 0)
 	{
-		((int32*)m_column_types.Items())[column_index] = CLVColNone;
+		((intptr_t*)m_column_types.Items())[column_index] = CLVColNone;
 		((char**)m_column_content.Items())[column_index] = NULL;
 		((char**)m_aux_content.Items())[column_index] = NULL;
 	}
@@ -132,12 +132,12 @@ void CLVEasyItem::SetColumnContent(int column_index, const char *text, bool trun
 
 		if(!truncate)
 		{
-			((int32*)m_column_types.Items())[column_index] = CLVColStaticText;
+			((intptr_t*)m_column_types.Items())[column_index] = CLVColStaticText;
 			((char**)m_aux_content.Items())[column_index] = NULL;
 		}
 		else
 		{
-			((int32*)m_column_types.Items())[column_index] = CLVColTruncateText|CLVColFlagNeedsTruncation;
+			((intptr_t*)m_column_types.Items())[column_index] = CLVColTruncateText|CLVColFlagNeedsTruncation;
 			copy = new char[strlen(text)+3];
 			strcpy(copy,text);
 			((char**)m_aux_content.Items())[column_index] = copy;
@@ -153,16 +153,16 @@ void CLVEasyItem::SetColumnContent(int column_index, const BBitmap *bitmap, floa
 	//Create the new entry
 	if(bitmap == NULL)
 	{
-		((int32*)m_column_types.Items())[column_index] = CLVColNone;
+		((intptr_t*)m_column_types.Items())[column_index] = CLVColNone;
 		((char**)m_column_content.Items())[column_index] = NULL;
 		((char**)m_aux_content.Items())[column_index] = NULL;
 	}
 	else
 	{
 		if(copy)
-			((int32*)m_column_types.Items())[column_index] = CLVColBitmap|CLVColFlagBitmapIsCopy;
+			((intptr_t*)m_column_types.Items())[column_index] = CLVColBitmap|CLVColFlagBitmapIsCopy;
 		else
-			((int32*)m_column_types.Items())[column_index] = CLVColBitmap;
+			((intptr_t*)m_column_types.Items())[column_index] = CLVColBitmap;
 		BBitmap* the_bitmap;
 		if(copy)
 		{
@@ -183,7 +183,7 @@ void CLVEasyItem::SetColumnContent(int column_index, const BBitmap *bitmap, floa
 
 const char* CLVEasyItem::GetColumnContentText(int column_index) const
 {
-	int32 type = ((int32)m_column_types.ItemAt(column_index)) & CLVColTypesMask;
+	intptr_t type = ((intptr_t)m_column_types.ItemAt(column_index)) & CLVColTypesMask;
 	if(type != CLVColStaticText && type != CLVColTruncateText)
 		return NULL;
 	return (const char*)m_column_content.ItemAt(column_index);
@@ -192,7 +192,7 @@ const char* CLVEasyItem::GetColumnContentText(int column_index) const
 
 const BBitmap* CLVEasyItem::GetColumnContentBitmap(int column_index) const
 {
-	int32 type = ((int32)m_column_types.ItemAt(column_index)) & CLVColTypesMask;
+	intptr_t type = ((intptr_t)m_column_types.ItemAt(column_index)) & CLVColTypesMask;
 	if(type != CLVColBitmap)
 		return NULL;
 	return (const BBitmap*)m_column_content.ItemAt(column_index);
@@ -200,7 +200,7 @@ const BBitmap* CLVEasyItem::GetColumnContentBitmap(int column_index) const
 
 float CLVEasyItem::GetColumnContentWidth(BView *owner, BFont *font, int column_index) const
 {
-	int32 type = ((int32)m_column_types.ItemAt(column_index)) & CLVColTypesMask;
+	intptr_t type = ((intptr_t)m_column_types.ItemAt(column_index)) & CLVColTypesMask;
 
 	switch(type & CLVColTypesMask) {
 		case CLVColBitmap:
@@ -248,7 +248,7 @@ void CLVEasyItem::DrawItemColumn(BView *owner, BRect item_column_rect, int32 col
 
 void CLVEasyItem::DrawItemColumnInner(BView *owner, BRect item_column_rect, int32 column_index, bool complete)
 {
-	int32 type = (int32)m_column_types.ItemAt(column_index);
+	intptr_t type = (intptr_t)m_column_types.ItemAt(column_index);
 	if(type == 0)
 		return;
 	bool needs_truncation = false;
@@ -268,7 +268,7 @@ void CLVEasyItem::DrawItemColumnInner(BView *owner, BRect item_column_rect, int3
 				BFont owner_font;
 				owner->GetFont(&owner_font);
 				TruncateText(column_index,item_column_rect.right-item_column_rect.left,&owner_font);
-				((int32*)m_column_types.Items())[column_index] = CLVColTruncateText;
+				((intptr_t*)m_column_types.Items())[column_index] = CLVColTruncateText;
 			}
 			text = (char*)m_aux_content.ItemAt(column_index);
 		}
@@ -280,7 +280,7 @@ void CLVEasyItem::DrawItemColumnInner(BView *owner, BRect item_column_rect, int3
 	{
 		const BBitmap* bitmap = (BBitmap*)m_column_content.ItemAt(column_index);
 		BRect bounds = bitmap->Bounds();
-		float horizontal_offset = (float)((int32)m_aux_content.ItemAt(column_index));
+		float horizontal_offset = (float)((intptr_t)m_aux_content.ItemAt(column_index));
 		item_column_rect.left += horizontal_offset;
 		item_column_rect.right = item_column_rect.left + (bounds.right-bounds.left);
 		item_column_rect.top += ceil(((item_column_rect.bottom-item_column_rect.top)-(bounds.bottom-bounds.top))/2.0);
@@ -312,8 +312,8 @@ int CLVEasyItem::CompareItems(const CLVListItem *a_Item1, const CLVListItem *a_I
 		Item2->m_column_types.CountItems() <= KeyColumn)
 		return 0;
 	
-	int32 type1 = ((int32)Item1->m_column_types.ItemAt(KeyColumn)) & CLVColTypesMask;
-	int32 type2 = ((int32)Item2->m_column_types.ItemAt(KeyColumn)) & CLVColTypesMask;
+	intptr_t type1 = ((intptr_t)Item1->m_column_types.ItemAt(KeyColumn)) & CLVColTypesMask;
+	intptr_t type2 = ((intptr_t)Item2->m_column_types.ItemAt(KeyColumn)) & CLVColTypesMask;
 
 	if(!((type1 == CLVColStaticText || type1 == CLVColTruncateText) &&
 		(type2 == CLVColStaticText || type2 == CLVColTruncateText)))
@@ -381,7 +381,7 @@ void CLVEasyItem::ColumnWidthChanged(int32 column_index, float column_width, Col
 				other_rect->OffsetBy(width_delta,0);
 		}
 
-	int32 type = ((int32)m_column_types.ItemAt(column_index)) & CLVColTypesMask;
+	intptr_t type = ((intptr_t)m_column_types.ItemAt(column_index)) & CLVColTypesMask;
 	if(type == CLVColTruncateText)
 	{
 		BRect bounds = the_view->Bounds();
@@ -391,13 +391,13 @@ void CLVEasyItem::ColumnWidthChanged(int32 column_index, float column_width, Col
 			BFont view_font;
 			the_view->GetFont(&view_font);
 			BRect invalid = TruncateText(column_index, column_width,&view_font);
-			((int32*)m_column_types.Items())[column_index] = CLVColTruncateText;
+			((intptr_t*)m_column_types.Items())[column_index] = CLVColTruncateText;
 			if(invalid != BRect(-1.0,-1.0,-1.0,-1.0))
 				the_view->Invalidate(invalid);
 		}
 		else
 			//If it's not onscreen flag it for truncation the next time it's drawn
-			((int32*)m_column_types.Items())[column_index] = CLVColTruncateText|CLVColFlagNeedsTruncation;
+			((intptr_t*)m_column_types.Items())[column_index] = CLVColTruncateText|CLVColFlagNeedsTruncation;
 	}
 }
 
@@ -407,7 +407,7 @@ void CLVEasyItem::FrameChanged(int32 column_index, BRect new_frame, ColumnListVi
 	BRect* cached_rect = (BRect*)m_cached_rects.ItemAt(column_index);
 	if(cached_rect == NULL)
 		return;
-	int32 type = ((int32)m_column_types.ItemAt(column_index)) & CLVColTypesMask;
+	intptr_t type = ((intptr_t)m_column_types.ItemAt(column_index)) & CLVColTypesMask;
 	if(type == CLVColTruncateText)
 		if(*cached_rect != new_frame)
 		{
@@ -419,13 +419,13 @@ void CLVEasyItem::FrameChanged(int32 column_index, BRect new_frame, ColumnListVi
 				BFont view_font;
 				the_view->GetFont(&view_font);
 				BRect invalid = TruncateText(column_index, new_frame.right-new_frame.left,&view_font);
-				((int32*)m_column_types.Items())[column_index] = CLVColTruncateText;
+				((intptr_t*)m_column_types.Items())[column_index] = CLVColTruncateText;
 				if(invalid != BRect(-1.0,-1.0,-1.0,-1.0))
 					the_view->Invalidate(invalid);
 			}
 			else
 			//If it's not onscreen flag it for truncation the next time it's drawn
-				((int32*)m_column_types.Items())[column_index] = CLVColTruncateText|CLVColFlagNeedsTruncation;
+				((intptr_t*)m_column_types.Items())[column_index] = CLVColTruncateText|CLVColFlagNeedsTruncation;
 		}
 }
